+++ src/pretalx/static/common/js/copy.js
@@ -1,10 +1,18 @@
 $(function() {
   $('[data-toggle="tooltip"]').tooltip()
-  $('[data-toggle="tooltip"]').click(function(e) {
+  $('[data-toggle="tooltip"], .copyable-text').click(function(e) {
     var $temp = $("<input>")
     $("body").append($temp)
-    $temp.val($(e.currentTarget).data("destination")).select()
+    const currentTarget = $(e.currentTarget)
+    $temp.val(currentTarget.data("destination")).select()
     document.execCommand("copy")
     $temp.remove()
+    const previousTitle = currentTarget.title
+    currentTarget.title = "Copied!"
+    currentTarget.tooltip('show')
+    window.setTimeout(400, () => {
+      currentTarget.title = previousTitle;
+      currentTarget.tooltip('hide')
+    })
   })
 })