+++ src/pretalx/submission/forms/submission.py
@@ -27,7 +27,9 @@ class InfoForm(RequestRequire, PublicContent, forms.ModelForm):
         instance = kwargs.get('instance')
         initial = kwargs.pop('initial', {})
         if not instance or not instance.submission_type:
-            initial['submission_type'] = self.event.cfp.default_type
+            initial['submission_type'] = getattr(self.access_code, 'submission_type', self.event.cfp.default_type)
+        if not instance and self.access_code:
+            initial['track'] = self.access_code.track
         if not instance or not instance.content_locale:
             initial['content_locale'] = self.event.locale
 
@@ -55,34 +57,35 @@ class InfoForm(RequestRequire, PublicContent, forms.ModelForm):
                 if not access_code or not access_code.track:
                     self.fields['track'].queryset = self.event.tracks.filter(requires_access_code=False)
                 else:
-                    self.fields['track'].queryset = self.event.tracks.filter(Q(requires_access_code=False) | Q(pk=access_code.pk))
+                    self.fields['track'].queryset = self.event.tracks.filter(Q(requires_access_code=False) | Q(pk=access_code.track.pk))
             elif not self.event.settings.use_tracks or instance and instance.state != SubmissionStates.SUBMITTED:
                 self.fields.pop('track')
 
     def _set_submission_types(self, instance=None):
         _now = now()
         if instance and instance.pk and (
-            instance.state != SubmissionStates.SUBMITTED or not self.event.cfp.is_open
+            instance.state != SubmissionStates.SUBMITTED
+            or not self.event.cfp.is_open
         ):
             self.fields['submission_type'].queryset = self.event.submission_types.filter(pk=instance.submission_type_id)
             self.fields['submission_type'].disabled = True
             return
         access_code = self.access_code or getattr(instance, 'access_code', None)
-        if access_code and access_code.submission_type:
-            queryset = self.event.submission_types.filter(
-                Q(requires_access_code=False) | Q(pk=self.submission_type.pk)
-            )
+        if access_code and not access_code.submission_type:
+            pks = self.event.submission_types.values_list('pk', flat=True)
         else:
             queryset = self.event.submission_types.filter(requires_access_code=False)
-        if (
-            not self.event.cfp.deadline or self.event.cfp.deadline >= _now
-        ):  # No global deadline or still open
-            types = queryset.exclude(deadline__lt=_now)
-        else:
-            types = queryset.filter(deadline__gte=_now)
-        pks = set(types.values_list('pk', flat=True))
-        if instance and instance.pk:
-            pks |= {instance.submission_type.pk}
+            if (
+                not self.event.cfp.deadline or self.event.cfp.deadline >= _now
+            ):  # No global deadline or still open
+                types = queryset.exclude(deadline__lt=_now)
+            else:
+                types = queryset.filter(deadline__gte=_now)
+            pks = set(types.values_list('pk', flat=True))
+            if access_code:
+                pks = pks | {access_code.submission_type.pk}
+            if instance and instance.pk:
+                pks |= {instance.submission_type.pk}
         self.fields['submission_type'].queryset = self.event.submission_types.filter(
             pk__in=pks
         )