+++ src/pretalx/submission/models/cfp.py
@@ -55,6 +55,8 @@ class CfP(LogMixin, models.Model):
         new_type = '{types}new'
         tracks = '{base}tracks/'
         new_track = '{tracks}new'
+        access_codes = '{base}access-codes/'
+        new_access_code = '{access_codes}new'
         public = '{self.event.urls.base}cfp'
         submit = '{self.event.urls.base}submit/'